feat: Improve blockdim optimization script substantially
This commit makes the process of generating the optimal blockdims more
convenient by not having to manually change the number of GPUs included.
Crucially, it exits the process if an error is detected. Previously,
manual detection of errors was required.

TODO: This commit does not optimize deep halos yet. It only works for
HALO_DEPTH=1