fix: Handle results generation properly
Previously, the result outputs were unstable as the did not include the
cudaMalloc part of the code. This is a significant part of the
implementation, and should of course be included.

Also, the results can now be generated without checking for round mean
square errors. This speeds up the process significantly.