feat: Remove shared memory ghost zone fetching
Previously, the implementation padded the shared memory regions with a
ghost zone, adding significant overhead to the solution. The ghost zone
were not reused by other stencils within the same thread block (this is
only true for stencils looking like a cross, but is not true for, e.g.,
square stencils), and therefore imposed unecessary overhead. Also, this
caused trouble with restrictions on the amount of shared memory
available. This feature will be added later when we introduce square
stencils.