feat: Reduce register pressure by using smaller data type
This commit fixes a very peculiar bug. The smem kernel was unable to
launch with 1024 threads for STENCIL_DEPTH=32 because the register
pressure was too high. However, it works for larger stencil depths,
which explains its peculiarness. There must be some unnusual compiler
optimizations for this stencil depth. The commit does not edit the
other kernels, as using smaller data types might increase instructions
needed to perform the arithmetic.