feat: Resume autotuning from intermediate save
As the autotuning process takes a significant amount of time, we
instruct the process to save the autotuning progress intermediately
while progressing. If the script is somehow cancelled (e.g., SLURM
timeout), then when the script is reinitialized it can continue from
the previous save, saving a significant amount of time.